Separated Checker from CheckerPage
